feat(frontend): apply new confirmation modal

+++ frontend/src/app/admin/accounts/list.component.ts
@@ -2,12 +2,17 @@
 import { first } from 'rxjs/operators';
 
 import { AccountService } from '@app/_services';
+import { ConfirmationModal } from '@app/_components/confirmation.component';
+import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 
 @Component({ templateUrl: 'list.component.html' })
 export class ListComponent implements OnInit {
     accounts?: any[];
 
-    constructor(private accountService: AccountService) { }
+    constructor(
+        private accountService: AccountService,
+        private modalService: NgbModal
+    ) { }
 
     ngOnInit() {
         this.accountService.getAll()
@@ -16,12 +21,19 @@ export class ListComponent implements OnInit {
     }
 
     deleteAccount(id: string) {
-        const account = this.accounts!.find(x => x.id === id);
-        account.isDeleting = true;
-        this.accountService.delete(id)
-            .pipe(first())
-            .subscribe(() => {
-                this.accounts = this.accounts!.filter(x => x.id !== id)
-            });
+        let ref = this.modalService.open(ConfirmationModal);
+        ref.componentInstance.buildModal("deletion");
+        ref.componentInstance.bodyText = "Are you sure that you want to delete this account?";
+        ref.result.then((result) => {
+            if (result === 'ok click') {
+                const account = this.accounts!.find(x => x.id === id);
+                account.isDeleting = true;
+                this.accountService.delete(id)
+                    .pipe(first())
+                    .subscribe(() => {
+                        this.accounts = this.accounts!.filter(x => x.id !== id)
+                    });
+            }
+        });
     }
 }
